+++ internal/topic/topicreaderinternal/stream_reconnector_test.go
@@ -11,6 +11,7 @@ import (
 
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/background"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/empty"
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/topic"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/topic/topicreadercommon"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/value"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xcontext"
@@ -65,6 +66,9 @@ func TestTopicReaderReconnectorReadMessageBatch(t *testing.T) {
 
 				return baseReader, nil
 			},
+			retrySettings: topic.RetrySettings{
+				StartTimeout: topic.DefaultStartTimeout,
+			},
 			streamErr: errUnconnected,
 			tracer:    &trace.Topic{},
 		}
@@ -108,6 +112,9 @@ func TestTopicReaderReconnectorReadMessageBatch(t *testing.T) {
 				connectCalled++
 
 				return readers[connectCalled-1], nil
+			},
+			retrySettings: topic.RetrySettings{
+				StartTimeout: topic.DefaultStartTimeout,
 			},
 			streamErr: errUnconnected,
 			tracer:    &trace.Topic{},
@@ -228,9 +235,12 @@ func TestTopicReaderReconnectorConnectionLoop(t *testing.T) {
 		newStream2.EXPECT().CloseWithError(gomock.Any(), gomock.Any()).MinTimes(1)
 
 		reconnector := &readerReconnector{
-			connectTimeout: value.InfiniteDuration,
-			background:     *background.NewWorker(ctx, "test-worker, "+t.Name()),
+			background: *background.NewWorker(ctx, "test-worker, "+t.Name()),
+			retrySettings: topic.RetrySettings{
+				StartTimeout: value.InfiniteDuration,
+			},
 			tracer:         &trace.Topic{},
+			connectTimeout: value.InfiniteDuration,
 		}
 		reconnector.initChannelsAndClock()
 
@@ -499,7 +509,7 @@ func TestTopicReaderReconnectorReconnectWithError(t *testing.T) {
 		tracer:    &trace.Topic{},
 	}
 	reconnector.initChannelsAndClock()
-	err := reconnector.reconnect(ctx, nil, nil)
+	err := reconnector.reconnect(ctx, errors.New("test-reconnect"), nil)
 	require.ErrorIs(t, err, testErr)
 	require.ErrorIs(t, reconnector.streamErr, testErr)
 }