+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Fixed handle reconnection timeout error
 * Fixed experimental topic listener handle stop partition event
 
 ## v3.76.1